refactor: consolidate own-property helper and tighten __proto__ regression tests
