+++ packages/toon/src/encode/replacer.ts
@@ -1,5 +1,6 @@
 import type { EncodeReplacer, JsonArray, JsonObject, JsonValue } from '../types.ts'
-import { isJsonArray, isJsonObject, normalizeValue, setOwnProperty } from './normalize.ts'
+import { setOwnProperty } from '../shared/object-utils.ts'
+import { isJsonArray, isJsonObject, normalizeValue } from './normalize.ts'
 
 /**
  * Applies a replacer function to a `JsonValue` and all its descendants.
