+++ packages/toon/test/encode-security.test.ts
@@ -22,24 +22,6 @@ describe('encode security hardening', () => {
     }
   })
 
-  it('preserves own __proto__ properties when applying a replacer', () => {
-    const marker = '__toonReplacerPolluted'
-    const input = JSON.parse(`{"__proto__":{"${marker}":true},"b":2}`)
-    const identityReplacer: EncodeReplacer = (_key, value) => value
-
-    try {
-      const decoded = decode(encode(input, { replacer: identityReplacer })) as Record<string, any>
-
-      expect(Object.hasOwn(decoded, prototypeKey)).toBe(true)
-      expect(decoded[prototypeKey][marker]).toBe(true)
-      expect(decoded.b).toBe(2)
-      expect(({} as Record<string, unknown>)[marker]).toBeUndefined()
-    }
-    finally {
-      delete (Object.prototype as Record<string, unknown>)[marker]
-    }
-  })
-
   it('preserves nested own __proto__ properties introduced by a replacer', () => {
     const marker = '__toonNestedReplacerPolluted'
     const replacement = JSON.parse(`{"__proto__":{"${marker}":true},"safe":1}`)
